+++ src/byteShard/Internal/Authentication/Provider/Ldap.php
@@ -2,13 +2,14 @@
 
 namespace byteShard\Internal\Authentication\Provider;
 
+use byteShard\Authentication\User;
 use byteShard\Internal\Authentication\Authentication;
 use byteShard\Internal\Authentication\AuthenticationAction;
 use byteShard\Internal\Authentication\LdapProviderInterface;
 use byteShard\Internal\Authentication\ProviderInterface;
+use byteShard\Internal\Authentication\Providers;
 use byteShard\Internal\Login\Struct\Credentials;
-use byteShard\Ldap\Attribute;
-use byteShard\Ldap\Attributes;
+use byteShard\Ldap\Enum\ResultObject;
 use byteShard\Ldap\Filter;
 use byteShard\Session;
 use config;
@@ -18,8 +19,8 @@ class Ldap implements ProviderInterface
     private string $username;
 
     public function __construct(
-        private readonly ?LdapProviderInterface $authenticationObject = null)
-    {
+        private readonly ?LdapProviderInterface $authenticationObject = null
+    ) {
     }
 
     public function authenticate(?Credentials $credentials = null): bool
@@ -52,11 +53,32 @@ class Ldap implements ProviderInterface
 
             $filter = new Filter($config->getLdapBaseDn());
             $filter->setFilter($config->getLdapUid().'='.$loginUsername);
-            $users = $ldap->getArray($filter, new Attributes(new Attribute('dn')));
+            $users = $ldap->getArray(
+                $filter,
+                $config->getLdapAttributes(),
+            );
             if (count($users) === 1) {
-                $credentials->setUsername($users[0]->dn);
+                $user = $users[0];
+                $credentials->setUsername($user->dn);
                 $authenticated = $this->getLdapInstance($ldapHost, $ldapPort, $ldapMethod)->authenticate($credentials);
                 if ($authenticated === true) {
+                    $userObject = User::createUser(
+                        username: $user->{ResultObject::Username->value} ?? '',
+                        firstname: $user->{ResultObject::Firstname->value} ?? '',
+                        lastname: $user->{ResultObject::Lastname->value} ?? '',
+                        mail: $user->{ResultObject::Mail->value} ?? '',
+                        provider: Providers::LDAP
+                    );
+                    if (isset($user->{ResultObject::Groups->value})) {
+                        if (is_array($user->{ResultObject::Groups->value})) {
+                            foreach ($user->{ResultObject::Groups->value} as $group) {
+                                $userObject->addGroup($this->parseGroupName($group));
+                            }
+                        } else {
+                            $userObject->addGroup($this->parseGroupName($user->{ResultObject::Groups->value}));
+                        }
+                    }
+                    $userObject->store();
                     $this->username = $loginUsername;
                     return true;
                 }
@@ -65,6 +87,14 @@ class Ldap implements ProviderInterface
             Authentication::logout(action: AuthenticationAction::INVALID_CREDENTIALS);
         }
         Authentication::logout();
+    }
+
+    private function parseGroupName(string $groupName): string
+    {
+        if (preg_match('/cn=([^,]+)/i', $groupName, $matches)) {
+            return $matches[1];
+        }
+        return $groupName;
     }
 
     private function getLdapInstance(string $host, int $port = 389, string $method = ''): \byteShard\Ldap
@@ -107,6 +137,6 @@ class Ldap implements ProviderInterface
 
     public function logout(): void
     {
-
+        User::logout();
     }
 }